@@ -1,6 +1,6 @@
 {
-    "name": "@akanass/otpjs",
-    "version": "2.0.0",
+    "name": "@akanass/rx-otp",
+    "version": "1.0.0",
     "description": "HMAC-based (HOTP) and Time-based (TOTP) One-Time Password manager. Works with Google Authenticator for Two-Factor Authentication.",
     "main": "commonjs/index.js",
     "module": "esm5/index.js",
@@ -8,14 +8,13 @@
     "types": "index.d.ts",
     "private": false,
     "scripts": {
-        "pretest": "make pretest",
+        "pretest": "make clean && make pretest",
         "test": "make test",
         "coveralls": "make coveralls",
         "clean": "make clean",
-        "tsc": "make tsc",
         "packaging": "make packaging",
         "browserify": "make browserify",
-        "prebuild": "make test && make clean",
+        "prebuild": "make clean && make pretest && make test",
         "build": "make commonjs && make esm2015 && make esm5",
         "postbuild": "make packaging && make browserify"
     },
@@ -60,37 +59,34 @@
     },
     "homepage": "https://github.com/njl07/otp.js#readme",
     "dependencies": {
-        "@types/joi": "^13.0.5",
-        "@types/node": "^9.4.6",
+        "@types/node": "^10.12.19",
         "@types/qr-image": "^3.2.1",
-        "@types/uid-safe": "^2.1.1",
-        "big-integer": "^1.6.26",
-        "binstring": "^0.2.1",
+        "@types/uid-safe": "^2.1.2",
+        "ajv": "^6.7.0",
+        "big-integer": "^1.6.41",
         "convert-base": "^0.1.0",
-        "joi": "^13.1.2",
         "pad-component": "^0.0.1",
         "qr-image": "^3.2.0",
         "thirty-two": "^1.0.2",
         "uid-safe": "^2.1.5"
     },
     "devDependencies": {
-        "@types/fs-extra": "^5.0.0",
-        "browserify": "^16.1.0",
-        "coveralls": "^3.0.0",
+        "@types/fs-extra": "^5.0.4",
+        "@types/jest": "^23.3.13",
+        "browserify": "^16.2.3",
+        "coveralls": "^3.0.2",
         "exorcist": "^1.0.1",
-        "fs-extra": "^5.0.0",
-        "istanbul": "^1.1.0-alpha.1",
-        "mocha": "^5.0.1",
-        "mocha-typescript": "^1.1.12",
-        "rimraf": "^2.6.2",
-        "rxjs": "^5.5.6",
-        "ts-node": "^5.0.0",
-        "tslint": "^5.9.1",
-        "typescript": "^2.7.2",
-        "unit.js": "^2.0.0"
+        "fs-extra": "^7.0.1",
+        "jest": "~23.6.0",
+        "rimraf": "^2.6.3",
+        "rxjs": "^6.4.0",
+        "ts-jest": "^23.10.5",
+        "ts-node": "^8.0.2",
+        "tslint": "^5.12.1",
+        "typescript": "^3.2.4"
     },
     "peerDependencies": {
-        "rxjs": "^5.5.6"
+        "rxjs": "^6.4.0"
     },
     "engines": {
         "node": ">=7.0.0"
